Re-config: Re-config chunk
Relates to #4
