+++ chunk_init_common.go
@@ -83,7 +83,10 @@ func (i *chunkInitCommon) unmarshal(raw []byte) error {
 	remaining := len(raw) - offset
 	for remaining > 0 {
 		if remaining > initOptionalVarHeaderLength {
-			pType := paramType(binary.BigEndian.Uint16(raw[offset:]))
+			pType, err := parseParamType(raw[offset:])
+			if err != nil {
+				return errors.Wrap(err, "failed to parse param type")
+			}
 			p, err := buildParam(pType, raw[offset:])
 			if err != nil {
 				return errors.Wrap(err, "Failed unmarshalling param in Init Chunk")
@@ -123,8 +126,7 @@ func (i *chunkInitCommon) marshal() ([]byte, error) {
 		// parameter except the last parameter in the chunk.*  The receiver
 		// MUST ignore the padding.
 		if idx != len(i.params)-1 {
-			padding := make([]byte, getPadding(len(pp)))
-			out = append(out, padding...)
+			out = padByte(out, getPadding(len(pp)))
 		}
 	}
