+++ param_outgoing_reset_request_test.go
@@ -6,29 +6,34 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
+var (
+	testChunkReconfigParamA = []byte{0x0, 0xd, 0x0, 0x16, 0x0, 0x0, 0x0, 0x1, 0x0, 0x0, 0x0, 0x2, 0x0, 0x0, 0x0, 0x3, 0x0, 0x4, 0x0, 0x5, 0x0, 0x6}
+	testChunkReconfigParamB = []byte{0x0, 0xd, 0x0, 0x10, 0x0, 0x0, 0x0, 0x1, 0x0, 0x0, 0x0, 0x2, 0x0, 0x0, 0x0, 0x3}
+)
+
 func TestParamOutgoingResetRequest_Success(t *testing.T) {
 	tt := []struct {
 		binary []byte
 		parsed *paramOutgoingResetRequest
 	}{
-		{[]byte{0x0, 0xd, 0x0, 0x14, 0x0, 0x0, 0x0, 0x1, 0x0, 0x0, 0x0, 0x2, 0x0, 0x0, 0x0, 0x3, 0x0, 0x4, 0x0, 0x5},
+		{testChunkReconfigParamA,
 			&paramOutgoingResetRequest{
 				paramHeader: paramHeader{
 					typ: outSSNResetReq,
-					len: 20,
-					raw: []byte{0x0, 0x0, 0x0, 0x1, 0x0, 0x0, 0x0, 0x2, 0x0, 0x0, 0x0, 0x3, 0x0, 0x4, 0x0, 0x5},
+					len: 22,
+					raw: testChunkReconfigParamA[4:],
 				},
 				reconfigRequestSequenceNumber:  1,
 				reconfigResponseSequenceNumber: 2,
 				senderLastTSN:                  3,
-				streamIdentifiers:              []uint16{4, 5},
+				streamIdentifiers:              []uint16{4, 5, 6},
 			}},
-		{[]byte{0x0, 0xd, 0x0, 0x10, 0x0, 0x0, 0x0, 0x1, 0x0, 0x0, 0x0, 0x2, 0x0, 0x0, 0x0, 0x3},
+		{testChunkReconfigParamB,
 			&paramOutgoingResetRequest{
 				paramHeader: paramHeader{
 					typ: outSSNResetReq,
 					len: 16,
-					raw: []byte{0x0, 0x0, 0x0, 0x1, 0x0, 0x0, 0x0, 0x2, 0x0, 0x0, 0x0, 0x3},
+					raw: testChunkReconfigParamB[4:],
 				},
 				reconfigRequestSequenceNumber:  1,
 				reconfigResponseSequenceNumber: 2,
@@ -48,7 +53,7 @@ func TestParamOutgoingResetRequest_Success(t *testing.T) {
 		if err != nil {
 			t.Fatalf("failed to marshal: %v", err)
 		}
-		assert.Equal(t, b, tc.binary)
+		assert.Equal(t, tc.binary, b)
 	}
 }
 
@@ -57,7 +62,8 @@ func TestParamOutgoingResetRequest_Failure(t *testing.T) {
 		name   string
 		binary []byte
 	}{
-		{"packet too short", []byte{0x0, 0xd, 0x0, 0x14, 0x0, 0x0, 0x0, 0x1}},
+		{"packet too short", testChunkReconfigParamA[:8]},
+		{"param too short", []byte{0x0, 0xd, 0x0, 0x4}},
 	}
 
 	for i, tc := range tt {
