+++ paramheader_test.go
@@ -6,12 +6,16 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
+var (
+	testParamHeader = []byte{0x0, 0x1, 0x0, 0x4}
+)
+
 func TestParamHeader_Success(t *testing.T) {
 	tt := []struct {
 		binary []byte
 		parsed *paramHeader
 	}{
-		{[]byte{0x0, 0x1, 0x0, 0x4},
+		{testParamHeader,
 			&paramHeader{
 				typ: heartbeatInfo,
 				len: 4,
@@ -30,7 +34,7 @@ func TestParamHeader_Success(t *testing.T) {
 		if err != nil {
 			t.Errorf("failed to marshal: %v", err)
 		}
-		assert.Equal(t, b, tc.binary)
+		assert.Equal(t, tc.binary, b)
 	}
 }
 
@@ -39,8 +43,10 @@ func TestParamHeaderUnmarshal_Failure(t *testing.T) {
 		name   string
 		binary []byte
 	}{
-		{"header too short", []byte{0x0, 0x1}},
-		{"wrong reported length", []byte{0x0, 0x1, 0x0, 0x10}},
+		{"header too short", testParamHeader[:2]},
+		// {"wrong param type", []byte{0x0, 0x0, 0x0, 0x4}}, // Not possible to fail parseParamType atm.
+		{"reported length below header length", []byte{0x0, 0xd, 0x0, 0x3}},
+		{"wrong reported length", testChunkReconfigParamA[:4]},
 	}
 
 	for i, tc := range tt {
